Added graph versions and moved the status on right side.
