+++ src/actions/actionTypes.js
@@ -166,6 +166,7 @@ export const SUCCESS_TRANSFER_DATA = 'SUCCESS_TRANSFER_DATA';
 export const ERROR_TRANSFER_DATA = 'ERROR_TRANSFER_DATA';
 
 export const RECEIVED_GRAPH_STATUS = 'RECEIVED_GRAPH_STATUS';
+export const RECEIVED_GRAPH_VERSIONS = 'RECEIVED_GRAPH_VERSIONS';
 
 export const ERRORS_SUPPLIERS = 'ERRORS_SUPPLIERS';
